jwt token, post access types, small improvements

+++ src/user/service/serviceMethods/delete.ts
@@ -1,4 +1,11 @@
 import { Mongo } from "../../../mongo";
-export const deleteUser = async (email: string, password: string) => {
+import { checkCredentials } from "../../../utils/checkCredentials";
+import { validateAuthTokenSignature } from "../../../utils/validateAuthTokenSignature";
+
+export const deleteUser = async (token: string) => {
+  const { email, password } = validateAuthTokenSignature(token);
+  const user = await checkCredentials(email, password);
+  if (!user) throw new Error("Invalid credentials");
+  await Mongo.profiles().deleteOne({ userId: String(user._id) });
   return await Mongo.users().deleteOne({ email, password });
 };
